refactor(writer-step3): mechanically split writer behind stable facade
